requirements.txt
# SHM Decision-Support System dependencies
# Python 3.11+

# Core scientific stack
numpy>=1.26,<2.0
scipy>=1.12
pandas>=2.2

# Machine learning
scikit-learn>=1.4
torch>=2.2

# API
fastapi>=0.111
uvicorn[standard]>=0.29
pydantic>=2.7
httpx>=0.27

# Dashboard
streamlit>=1.35
plotly>=5.21
requests>=2.32
matplotlib>=3.8

# Config
pyyaml>=6.0

# Testing
pytest>=8.0
pytest-cov>=5.0

# Optional experiment tracking
# wandb>=0.17
